+++ T03/app/src/main/java/ca/ualberta/cmput301/t03/datamanager/jobs/WriteDataJob.java
@@ -1,4 +1,4 @@
-package ca.ualberta.cmput301.t03.datamanager.queueddatamanager;
+package ca.ualberta.cmput301.t03.datamanager.jobs;
 
 
 import java.io.IOException;
